+++ man-roxygen/constraintParam.R
@@ -1,4 +1,6 @@
-#' @param constraint An object of class `phyDat`; returned trees will be
-#' perfectly compatible with each character in `constraint`.
+#' @param constraint Either an object of class `phyDat`, in which case returned
+#' trees will be perfectly compatible with each character in `constraint`;
+#' or a tree of class `phylo`, in which each node in `constraint` will occur in
+#' the returned tree.
 #' See [vignette](https://ms609.github.io/TreeSearch/articles/tree-search.html)
 #' for further examples.